+++ app/src/main/java/com/project/cinenook/data/MoviePopularResponse.kt
@@ -47,10 +47,10 @@ data class ResultsItem(
     val releaseDate: String? = null,
 
     @field:SerializedName("popularity")
-    val popularity: Double? = null,
+    val popularity: Float? = null,
 
     @field:SerializedName("vote_average")
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
 
     @field:SerializedName("id")
     val id: Int? = null,
